refactor: split base, hook and task into their own files
